requirements.txt
# Core dependencies
opencv-python>=4.8.0
requests>=2.31.0
python-dotenv>=1.0.0

# Web server (admin panel + proxy)
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
python-multipart>=0.0.6
